ThinningAndShift

Warning

The documentation is still missing. Please see the "examples" section for usage.

Examples

First example

Some data I needed to generate

ThinningAndShift.GTAS

ThinningAndShift.Jittering

ThinningAndShift.make_poisson_samples

ThinningAndShift.make_samples_with_parent

ThinningAndShift.modulated_event_train

ThinningAndShift.self_interacting_1D_train

ThinningAndShift.GTAS — Type

Generalized Thinning & Shift (GTAS) generative model for spike-trains.

Parameters

n::I: number of emitting units

rate_parent::R: rate of the parent Poisson process, upper bound for the instantaneous rate of all units

markings::Vector{Vector{I}}: markings of the parent Poisson process, each element is a vector of unit indices that are marked by the parent Poisson process

marking_probs::Vector{R}: probabilities of each marking

jitters::Vector{J}: jitters for each marking

ThinningAndShift.Jittering — Type

Jittering selects the type of perturbation applied to the event times.

NoJitter: no jittering

JitterDistribution(Distributions.UnivariateDistribution): Sample i.i.d. and add to event times JitterIncremental(Distributions.UnivariateDistribution): apply a distribution to each spike time, cumulatively JitterPaired(Distributions.UnivariateDistribution): apply a distribution to each pair of spike times

ThinningAndShift.make_poisson_samples — Method

makepoissonsamples(rate::Real, t_tot::Real)

Sample from a uniform Poisson process with rate rate on the interval [0, t_tot]. Returns a vector of event times of variable size.

ThinningAndShift.make_samples_with_parent — Method

makesampleswithparent(g::GTAS{R,I}, ttot::R) where {R,I}

Samples from a generalized thinning and shift process, returning event times in the (0,t_tot) interval.

Parameters:

g::GTAS{R,I} : the generalized thinning and shift process

t_tot::R : the total time

Returns:

trains::Vector{Vector{R}} : the trains of events for each unit

ts_ancestor::Vector{R} : the ancestor train

attributions::Vector{I} : the marking attributed to each event in the ancestor train (therefore it has the same length)

ThinningAndShift.modulated_event_train — Method

modulatedeventtrain(ratefunc::Function, ratemax::Real, ttot::Real) -> modulatedtrain::Vector{Float64}

Generates a Poisson train with instantaneous rate matching rate_function(t) for t in [0,t_tot].

Based on simple thinning: samples from a Poisson process with rate rate_max and keeps each sample with probability rate_func(t) / rate_max.

rate_max must be strictly greater than rate_function(t) for all t in [0,t_tot].

Parameters:

rate_func::Function : the rate function

rate_max::Real : the maximum rate

t_tot::Real : the total time

ThinningAndShift.self_interacting_1D_train — Method

function selfinteracting1Dtrain(kernelfunc::Function, ratestart::Real, ttot::Real; ratemax::Union{Real,Nothing}=nothing, verbose::Bool=false, timehorizon::Union{Real,Nothing}=nothing)

Generates a 1D self-interacting Poisson process following an arbitrary interaction kernel function.

IMPORTANT: the kernel function cannot have area larger than 1. And rate_max must be strictly greater than rate_function(t) for all t in [0,t_tot]. this is non-trivial, as repeated nearby spikes result in multiple staked kernels.

Final rate will be ratestart/(1-kernelarea)

Parameters:

kernel_func::Function : the interaction kernel function

rate_start::Real : the initial rate of the process

t_tot::Real : the total time of the process

rate_max::Union{Real} : the maximum rate of the process

verbose::Bool, default false : whether to print progress

time_horizon::Union{Real,Nothing}, default nothing : the time horizon for the kernel function. If provided, it optimizes the generation process by only considering spikes within this time horizon.