ThinningAndShift

Warning

The documentation is still missing. Please see the "examples" section for usage.

Examples

First example

Some data I needed to generate

ThinningAndShift.Jittering

ThinningAndShift.make_poisson_samples

ThinningAndShift.modulated_event_train

ThinningAndShift.Jittering — Type

Jittering selects the type of perturbation applied to the event times.

NoJitter: no jittering

JitterDistribution(Distribution.UnivariateDistribution): Sample i.i.d. and add to event times

JitterIncremental: apply a distribution to each spike time, cumulatively JitterPaired: apply a distribution to each pair of spike times

ThinningAndShift.make_poisson_samples — Method

makepoissonsamples(rate::Real, t_tot::Real)

Sample from a uniform Poisson process with rate rate on the interval [0, t_tot]. Returns a vector of event times of variable size.

ThinningAndShift.modulated_event_train — Method

modulatedeventtrain(ratefunc::Function, ratemax::Real, t_tot::Real)

Sample from a Poisson process with rate rate_func(t) on the interval [0, t_tot].

Uses simple thinning: samples from a Poisson process with rate rate_max and keeps each sample with probability rate_func(t) / rate_max.